# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gen-IT API Documentation",
        "description": "API documentation for Gen-IT application",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/submissions": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get all submissions with pagination",
                "operationId": "1ab2c30900bcf0239d12dadb880c27d4",
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "_limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "_search",
                        "in": "query",
                        "description": "Search by title or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by submission type",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by submission status",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of submissions with pagination"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Store a newly created submission",
                "operationId": "86842154f1dffa008adb43a94e7c19a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "type",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "link": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Submission created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Error creating submission"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/submissions/{id}": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Update an existing submission",
                "description": "Update a submission with optional file and image uploads.",
                "operationId": "5a5cdb632c1a644be575baebebdee6d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Submission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "content": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "link": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "img": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission updated successfully"
                    },
                    "404": {
                        "description": "Submission not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Delete a submission",
                "operationId": "eda69911376172d6b69e32fe2842e482",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Submission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission deleted successfully"
                    },
                    "404": {
                        "description": "Submission not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/submissions/{id}/approve": {
            "patch": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Approve a submission",
                "operationId": "ef91dcfbc498322ce6e83303b28d80e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Submission ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission approved successfully"
                    },
                    "404": {
                        "description": "Submission not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}